root/requirements.txt
Flask
Flask-SQLAlchemy
requests
aws-wsgi
Werkzeug
MarkupSafe
boto3
Flask-CORS
Flask-Injector
aiohttp
psycopg2-binary
Flask-Cognito
PyJWT